+++ packages/js/src/plugins/rfqModule/helpers.ts
@@ -7,7 +7,7 @@ import {
 } from '@solana/web3.js';
 import * as Spl from '@solana/spl-token';
 import { Sha256 } from '@aws-crypto/sha256-js';
-import { Side, Leg, FixedSize, StoredRfqState } from '@convergence-rfq/rfq';
+import { Quote, Side, Leg } from '@convergence-rfq/rfq';
 import * as anchor from '@project-serum/anchor';
 import * as psyoptionsAmerican from '@mithraic-labs/psy-american';
 import { OptionMarketWithKey } from '@mithraic-labs/psy-american';
@@ -44,7 +44,6 @@ import { collateralMintCache } from '../collateralModule';
 import { Mint } from '../tokenModule';
 import {
   LegInstrument,
-  QuoteInstrument,
   getSerializedLegLength,
   getValidationAccounts,
   serializeAsLeg,
@@ -52,7 +51,7 @@ import {
 } from '../instrumentModule';
 import { SpotLegInstrument } from '../spotInstrumentModule';
 import type { Rfq, Response } from './models';
-import { LEG_MULTIPLIER_DECIMALS } from './constants';
+import { ABSOLUTE_PRICE_DECIMALS, LEG_MULTIPLIER_DECIMALS } from './constants';
 
 const { mintOptions } = instructions;
 
@@ -226,51 +225,92 @@ export const convertOverrideLegMultiplierBps = (
   return overrideLegMultiplierBps * Math.pow(10, 9);
 };
 
-export const convertFixedSizeInput = (
-  fixedSize: FixedSize,
-  quoteAsset: QuoteInstrument
-): FixedSize => {
-  if (fixedSize.__kind == 'BaseAsset') {
-    fixedSize.legsMultiplierBps = addDecimals(
-      Number(fixedSize.legsMultiplierBps),
-      LEG_MULTIPLIER_DECIMALS
+export const convertResponseOutput = (
+  response: Response,
+  quoteDecimals: number
+): Response => {
+  if (response.bid) {
+    const convertedPriceQuoteAmountBps = removeDecimals(
+      response.bid.priceQuote.amountBps,
+      quoteDecimals
     );
-  } else if (fixedSize.__kind == 'QuoteAsset') {
-    fixedSize.quoteAmount = addDecimals(
-      Number(fixedSize.quoteAmount),
-      quoteAsset.getDecimals()
+    const removedQuoteDecimals = new anchor.BN(convertedPriceQuoteAmountBps);
+    const removedAbsoluteDecimals = removeDecimals(
+      removedQuoteDecimals,
+      ABSOLUTE_PRICE_DECIMALS
     );
-  }
 
-  return fixedSize;
+    response.bid.priceQuote.amountBps = removedAbsoluteDecimals;
+
+    if (response.bid.__kind == 'Standard') {
+      const convertedLegsMultiplierBps = removeDecimals(
+        response.bid.legsMultiplierBps
+      );
+
+      response.bid.legsMultiplierBps = new anchor.BN(
+        convertedLegsMultiplierBps
+      );
+    }
+  }
+  if (response.ask) {
+    const convertedPriceQuoteAmountBps = removeDecimals(
+      response.ask.priceQuote.amountBps,
+      quoteDecimals
+    );
+    const removedQuoteDeciamls = new anchor.BN(convertedPriceQuoteAmountBps);
+    const removedAbsoluteDecimals = removeDecimals(
+      removedQuoteDeciamls,
+      ABSOLUTE_PRICE_DECIMALS
+    );
+
+    response.ask.priceQuote.amountBps = removedAbsoluteDecimals;
+
+    if (response.ask.__kind == 'Standard') {
+      const convertedLegsMultiplierBps = removeDecimals(
+        response.ask.legsMultiplierBps
+      );
+      response.ask.legsMultiplierBps = new anchor.BN(
+        convertedLegsMultiplierBps
+      );
+    }
+  }
+  return response;
 };
 
-export const convertRfqOutput = (
-  rfq: Rfq,
-  collateralMintDecimals: number
-): Rfq => {
-  rfq.nonResponseTakerCollateralLocked = removeDecimals(
-    rfq.nonResponseTakerCollateralLocked,
-    collateralMintDecimals
+const convertQuoteInput = (quote: Quote, quoteDecimals: number) => {
+  const convertedQuote = structuredClone(quote);
+  convertedQuote.priceQuote.amountBps = quote.priceQuote.amountBps;
+  // multiply before and then convert to anchor.BN
+  const priceQuoteWithDecimals = addDecimals(
+    Number(convertedQuote.priceQuote.amountBps),
+    quoteDecimals
   );
-  rfq.totalTakerCollateralLocked = removeDecimals(
-    rfq.totalTakerCollateralLocked,
-    collateralMintDecimals
+  const convertedPriceQuoteAmountBps = new anchor.BN(priceQuoteWithDecimals);
+
+  convertedQuote.priceQuote.amountBps = convertedPriceQuoteAmountBps.mul(
+    new anchor.BN(10).pow(new anchor.BN(ABSOLUTE_PRICE_DECIMALS))
   );
 
-  if (rfq.fixedSize.__kind == 'BaseAsset') {
-    rfq.fixedSize.legsMultiplierBps = removeDecimals(
-      rfq.fixedSize.legsMultiplierBps,
+  if (convertedQuote.__kind == 'Standard') {
+    // multiply before and then convert to anchor.BN
+    const priceQuoteWithDecimals = addDecimals(
+      Number(convertedQuote.legsMultiplierBps),
       LEG_MULTIPLIER_DECIMALS
     );
-  } else if (rfq.fixedSize.__kind == 'QuoteAsset') {
-    rfq.fixedSize.quoteAmount = removeDecimals(
-      rfq.fixedSize.quoteAmount,
-      rfq.quoteAsset.getDecimals()
-    );
+    const convertedLegsMultiplierBps = new anchor.BN(priceQuoteWithDecimals);
+    convertedQuote.legsMultiplierBps = convertedLegsMultiplierBps;
   }
+  return convertedQuote;
+};
 
-  return rfq;
+export const convertResponseInput = (
+  quoteDecimals: number,
+  bid?: Quote,
+  ask?: Quote
+) => {
+  const convertedBid = bid ? convertQuoteInput(bid, quoteDecimals) : undefined;
+  const convertedAsk = ask ? convertQuoteInput(ask, quoteDecimals) : undefined;
+  return { convertedBid, convertedAsk };
 };
 
 export const calculateExpectedLegsHash = (
@@ -1519,7 +1559,7 @@ export const createAccountsAndMintOptions = async (
 export const sortByActiveAndExpiry = (rfqs: Rfq[]) => {
   return rfqs
     .sort((a, b) => {
-      return b.state === StoredRfqState.Active ? 1 : -1;
+      return b.state === 'active' ? 1 : -1;
     })
     .sort((a, b) => {
       if (a.state === b.state) {