+++ packages/js/src/plugins/rfqModule/index.ts
@@ -11,5 +11,4 @@ export * from './operations';
 export * from './pdas';
 export * from './plugin';
 export * from './program';
-export * from './types';
 export * from './constants';