+++ packages/js/src/plugins/rfqModule/operations/createAndFinalizeRfqConstruction.ts
@@ -2,10 +2,9 @@ import { PublicKey } from '@solana/web3.js';
 import * as anchor from '@project-serum/anchor';
 
 import { SendAndConfirmTransactionResponse } from '../../rpcModule';
-import { OrderType, FixedSize } from '../types';
-import { assertRfq, Rfq } from '../models';
+import { assertRfq, FixedSize, Rfq } from '../models';
 
-import { convertFixedSizeInput, calculateExpectedLegsHash } from '../helpers';
+import { calculateExpectedLegsHash } from '../helpers';
 import {
   Operation,
   OperationHandler,
@@ -21,6 +20,7 @@ import {
   QuoteInstrument,
   toQuote,
 } from '../../../plugins/instrumentModule';
+import { OrderType } from '../models/OrderType';
 import { createRfqBuilder } from './createRfq';
 import { finalizeRfqConstructionBuilder } from './finalizeRfqConstruction';
 
@@ -110,7 +110,6 @@ export type CreateAndFinalizeRfqConstructionInput = {
    * Optional address of the Taker's collateral info account.
    *
    * @defaultValue `convergence.collateral().pdas().collateralInfo({ user: taker.publicKey })`
-   *
    */
   collateralInfo?: PublicKey;
 
@@ -165,7 +164,6 @@ export const createAndFinalizeRfqConstructionOperationHandler: OperationHandler<
         Math.floor(Date.now() / 1_000) - 10
       );
 
-      const convertedFixedSize = convertFixedSizeInput(fixedSize, quoteAsset);
       const expectedLegsHash = calculateExpectedLegsHash(instruments);
 
       const rfqPda = convergence
@@ -176,7 +174,7 @@ export const createAndFinalizeRfqConstructionOperationHandler: OperationHandler<
           legsHash: Buffer.from(expectedLegsHash),
           orderType,
           quoteAsset: toQuote(quoteAsset),
-          fixedSize: convertedFixedSize,
+          fixedSize,
           activeWindow,
           settlingWindow,
           recentTimestamp,
@@ -188,7 +186,7 @@ export const createAndFinalizeRfqConstructionOperationHandler: OperationHandler<
           ...operation.input,
           taker,
           rfq: rfqPda,
-          fixedSize: convertedFixedSize,
+          fixedSize,
           instruments,
           expectedLegsHash,
           recentTimestamp,